# frontend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "CoffeeTracker.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/beans": {
      "get": {
        "tags": [
          "Beans"
        ],
        "operationId": "GetBeans",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeUnavailable",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "country",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BeanSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Beans"
        ],
        "operationId": "CreateBean",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBeanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBeanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/beans/{id}": {
      "get": {
        "tags": [
          "Beans"
        ],
        "operationId": "GetBeanById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeanDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Beans"
        ],
        "operationId": "UpdateBean",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBeanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/beans/{id}/image": {
      "put": {
        "tags": [
          "Beans"
        ],
        "operationId": "UploadBeanImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "get": {
        "tags": [
          "Beans"
        ],
        "operationId": "GetBeanImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Beans"
        ],
        "operationId": "DeleteBeanImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/beans/{id}/availability": {
      "patch": {
        "tags": [
          "Beans"
        ],
        "operationId": "SetBeanAvailability",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBeanAvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/beans/parse-image": {
      "post": {
        "tags": [
          "Beans"
        ],
        "operationId": "ParseBeanImage",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "imageFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseBeanImageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/countries/{countryId}/beans": {
      "get": {
        "tags": [
          "Countries"
        ],
        "operationId": "GetCountryBeans",
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BeanSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "roasterId": {
            "type": "string",
            "format": "uuid"
          },
          "roasterName": {
            "type": "string",
            "nullable": true
          },
          "originType": {
            "$ref": "#/components/schemas/OriginType"
          },
          "originCountries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BeanOriginCountryDto"
            },
            "nullable": true
          },
          "variety": {
            "type": "string",
            "nullable": true
          },
          "processingMethod": {
            "type": "string",
            "nullable": true
          },
          "roastProfile": {
            "$ref": "#/components/schemas/RoastProfile"
          },
          "roastDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "altitude": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bagWeight": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "pricePerKg": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "flavorNotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlavorNoteDto"
            },
            "nullable": true
          },
          "hasImage": {
            "type": "boolean"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean"
          },
          "remainingQuantity": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "BeanOriginCountryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BeanSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "roasterName": {
            "type": "string",
            "nullable": true
          },
          "roastProfile": {
            "$ref": "#/components/schemas/RoastProfile"
          },
          "roastDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "bagWeight": {
            "type": "number",
            "format": "double"
          },
          "pricePerKg": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "hasImage": {
            "type": "boolean"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean"
          },
          "remainingQuantity": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "CreateBeanRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "roasterId": {
            "type": "string",
            "format": "uuid"
          },
          "originType": {
            "$ref": "#/components/schemas/OriginType"
          },
          "originCountryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "variety": {
            "type": "string",
            "nullable": true
          },
          "processingMethod": {
            "type": "string",
            "nullable": true
          },
          "roastProfile": {
            "$ref": "#/components/schemas/RoastProfile"
          },
          "roastDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "altitude": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bagWeight": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "flavorNoteNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateBeanResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "FlavorNoteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OriginType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "ParseBeanImageResponse": {
        "type": "object",
        "properties": {
          "beanName": {
            "type": "string",
            "nullable": true
          },
          "roasterId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "roasterName": {
            "type": "string",
            "nullable": true
          },
          "originType": {
            "$ref": "#/components/schemas/OriginType"
          },
          "originCountries": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "variety": {
            "type": "string",
            "nullable": true
          },
          "processingMethod": {
            "type": "string",
            "nullable": true
          },
          "roastProfile": {
            "$ref": "#/components/schemas/RoastProfile"
          },
          "roastDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "altitude": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bagWeight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "flavorNotes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "unmatchedReferences": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RoastProfile": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SetBeanAvailabilityRequest": {
        "type": "object",
        "properties": {
          "isAvailable": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBeanRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "roasterId": {
            "type": "string",
            "format": "uuid"
          },
          "originType": {
            "$ref": "#/components/schemas/OriginType"
          },
          "originCountryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "variety": {
            "type": "string",
            "nullable": true
          },
          "processingMethod": {
            "type": "string",
            "nullable": true
          },
          "roastProfile": {
            "$ref": "#/components/schemas/RoastProfile"
          },
          "roastDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "altitude": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bagWeight": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean"
          },
          "flavorNoteNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}